+++ CodeTree/꼬리잡기놀이/jeehwan_꼬리잡기놀이.py
@@ -261,138 +261,4 @@ for i in range(k):
 
 print(ans)
 
-'''
-bfs 풀이 
-
-from collections import deque
-
-dx = [0, 1, 0, -1]
-dy = [1, 0, -1, 0]
-
-def find_group(x, y):
-    q = deque([(x, y)])
-    locations = [(board[x][y], x, y)]
-    groups[x][y] = group_num
-
-    while q:
-        x, y = q.popleft()
-        for k in range(4):
-            nx, ny = x + dx[k], y + dy[k]
-            if 0 <= nx < n and 0 <= ny < n:
-                if board[nx][ny] and not groups[nx][ny]:
-                    if board[x][y] == 1 and board[nx][ny] == 3: continue
-                    groups[nx][ny] = group_num
-                    q.append((nx, ny))
-                    if board[nx][ny] != 4:
-                        locations.append((board[nx][ny], nx, ny))
-    return locations
-
-def move_team(num):
-    head, x, y = group_info[num][0]
-    tmp = []
-    flag = False
-    for k in range(4):
-        nx, ny = x + dx[k], y + dy[k]
-        if 0 <= nx < n and 0 <= ny < n:
-            if board[nx][ny] == 4:
-                board[nx][ny] = head
-                tmp.append((head, nx, ny))
-                flag = True
-                break
-    for i in range(len(group_info[num])):
-        if i == 0 and flag: continue
-        now, x, y = group_info[num][i]
-        before, bx, by = group_info[num][i - 1]
-        tmp.append((now, bx, by))
-        board[bx][by] = now
-        board[x][y] = 4
-
-    head, hx, hy = tmp[0]
-    tail, tx, ty = tmp[-1]
-    board[hx][hy] = 1
-    board[tx][ty] = 3
-
-    group_info[num] = tmp
-
-def find_arrow(now_time):
-    global ans
-    a, b = (now_time // n) % 4, now_time % n
-    if a == 0:
-        for j in range(n):
-            if board[b][j] and board[b][j] != 4:
-                group_number = groups[b][j]
-                for idx, (p, x, y) in enumerate(group_info[group_number]):
-                    if (x, y) == (b, j):
-                        ans += (idx + 1) ** 2
-                        break
-                group_info[group_number].reverse()
-                head, hx, hy = group_info[group_number][0]
-                tail, tx, ty = group_info[group_number][-1]
-                board[hx][hy], board[tx][ty] = board[tx][ty], board[hx][hy]
-                break
-    elif a == 1:
-        for i in range(n - 1, -1, -1):
-            if board[i][b] and board[i][b] != 4:
-                group_number = groups[i][b]
-                for idx, (p, x, y) in enumerate(group_info[group_number]):
-                    if (x, y) == (i, b):
-                        ans += (idx + 1) ** 2
-                        break
-                group_info[group_number].reverse()
-                head, hx, hy = group_info[group_number][0]
-                tail, tx, ty = group_info[group_number][-1]
-                board[hx][hy], board[tx][ty] = board[tx][ty], board[hx][hy]
-                break
-    elif a == 2:
-        for j in range(n - 1, -1, -1):
-            if board[n - 1 - b][j] and board[n - 1 - b][j] != 4:
-                group_number = groups[n - 1 - b][j]
-                for idx, (p, x, y) in enumerate(group_info[group_number]):
-                    if (x, y) == (n - 1 - b, j):
-                        ans += (idx + 1) ** 2
-                        break
-                group_info[group_number].reverse()
-                head, hx, hy = group_info[group_number][0]
-                tail, tx, ty = group_info[group_number][-1]
-                board[hx][hy], board[tx][ty] = board[tx][ty], board[hx][hy]
-                break
-    elif a == 3:
-        for i in range(n):
-            if board[i][n - 1 - b] and board[i][n - 1 - b] != 4:
-                group_number = groups[i][n - 1 - b]
-                for idx, (p, x, y) in enumerate(group_info[group_number]):
-                    if (x, y) == (i, n - 1 - b):
-                        ans += (idx + 1) ** 2
-                        break
-                group_info[group_number].reverse()
-                head, hx, hy = group_info[group_number][0]
-                tail, tx, ty = group_info[group_number][-1]
-                board[hx][hy], board[tx][ty] = board[tx][ty], board[hx][hy]
-                break
-
-if __name__ == '__main__':
-    n, m, k = map(int, input().split())
-    board = [list(map(int, input().split())) for _ in range(n)]
-    time, ans = 0, 0
-
-    groups = [[0] * n for _ in range(n)]
-    group_num = 1
-    group_info = [[] for _ in range(m + 1)]
-
-    for i in range(n):
-        for j in range(n):
-            if board[i][j] == 1:
-                group_info[group_num] = find_group(i, j)
-                group_num += 1
-
-    while time < k:
-        # 각 팀 한칸 이동
-        for i in range(1, group_num):
-            move_team(i)
-        # 턴 수에 따른 화살표 계산 + 점수 계산 + 위치 변경
-        find_arrow(time)
-        # 다음 라운드
-        time += 1
-    print(ans)
-'''
